Move prepareBeforeQuit() and createQuitLabel() to manager instead of events file
